src/main/java/com/switchfully/bram/eurder/instances/valueObjects/Address.java
package com.switchfully.bram.eurder.instances.valueObjects;

import java.util.Objects;

public class Address {
    private final String streetName;
    private final String houseNumber;
    private final String city;
    private final int postalCode;

    public Address(String streetName, String houseNumber, String city, int postalCode) {
        this.streetName = streetName;
        this.houseNumber = houseNumber;
        this.city = city;
        this.postalCode = postalCode;
    }

    public String getStreetName() {
        return streetName;
    }

    public String getHouseNumber() {
        return houseNumber;
    }

    public String getCity() {
        return city;
    }

    public int getPostalCode() {
        return postalCode;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Address address = (Address) o;
        return postalCode == address.postalCode &&
                Objects.equals(streetName, address.streetName) &&
                Objects.equals(houseNumber, address.houseNumber) &&
                Objects.equals(city, address.city);
    }

}
